==> EDA/data-preprocessing/04-Scaling-and-Normalization.ipynb <==
## **Scaling and Normalization**

- Robust Scaler 
- Log Scaler 

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np
import math 
from scipy import stats

**Data**

# Load the data
try:
    data = pd.read_csv('data_versions/03_feature_engineered_data.csv', index_col=False)
    print(f"Successfully loaded {len(data.columns)} features")
except Exception as e: 
    print(f"Error loading data: {e}")

**Analysis** 

# Drop the first misc column
data.drop(columns='Unnamed: 0', inplace=True)

*Functions*

# Define Scaling functions
from sklearn.preprocessing import RobustScaler

def robust_scale_features(df):
    """
    Scales the features of the DataFrame using the RobustScaler.
    Parameters:
        df (pd.DataFrame): The input DataFrame.
    Returns:
        pd.DataFrame: The DataFrame with scaled features.
    """
    scaler = RobustScaler()
    scaled_df = pd.DataFrame(scaler.fit_transform(df), columns=df.columns)
    return scaled_df


def log_transform_features(df):
    """
    Applies logarithmic transformation to the specified features in the DataFrame.
    Parameters:
        df (pd.DataFrame): The input DataFrame.
        feature_list (list): List of feature names to be log-transformed.
    Returns:
        pd.DataFrame: The DataFrame with log-transformed features.  
    """
    df = df.applymap(lambda x: np.log1p(x) if x >= 0 else 0)
    return df


# Separate the label and features before transformation
y = data.iloc[:, -1]
x = data.drop(columns='Label')

# Apply Robust scale scaling
robust_scaled_data = robust_scale_features(x)

# sample data after robust scaler transformation
robust_scaled_data.sample(5)

# merge target col 
robust_scaled_data['Label'] = y

# Apply log transformation
try:
    data_log_transformed = log_transform_features(x)
    print(f"Successfully log transformed!")
except Exception as e:
    print(f"Error {e}")
    
# sample data after log transformation
data_log_transformed.sample(5)

# merge the target column back to the main data
data_log_scaled = pd.concat([data_log_transformed, y], axis=1)

We are saving log transformed data as of now, might use robust scaled data later if required

**Save to csv**

# Save to csv 
data_log_scaled.to_csv('data_versions/04_log_scaled_data.csv', index=False)
